+++ test/unit/test_gut_yielding.gd
@@ -39,11 +39,13 @@ func before_each():
 	timer.set_wait_time(1)
 
 func test_can_yield_using_built_in_timer():
+	pass_test('should have seen a pause')
 	gut.p('yielding for 1 second')
 	yield(gut.set_yield_time(1), 'timeout')
 	gut.p('done yielding')
 
 func test_setting_yield_time_twice_resets_time():
+	pass_test('should have seen a pause')
 	gut.p('yielding for 1 second')
 	gut.set_yield_time(10)
 	gut.set_yield_time(1)
@@ -51,14 +53,18 @@ func test_setting_yield_time_twice_resets_time():
 	gut.p('done yielding')
 
 func test_wait_for_continue_click():
+	gut.p('should have had to press continue')
 	assert_eq(1, 1, 'some simple assert')
 	gut.pause_before_teardown()
 
 func test_can_pause_twice():
+	gut.p('should have had to press continue')
 	assert_eq(2, 2, 'Another simple assert')
 	gut.pause_before_teardown()
 
+
 func test_will_wait_when_yielding():
+	pass_test('should have seen a pause')
 	timer.set_wait_time(5)
 	gut.p('yielding for 5 seconds')
 	timer.start()
@@ -66,6 +72,7 @@ func test_will_wait_when_yielding():
 	gut.p('done yielding')
 
 func test_can_pause_after_yielding():
+	pass_test('should have seen a pause and press continue')
 	gut.p('yielding for 1 second')
 	timer.start()
 	yield(timer, 'timeout')
@@ -73,6 +80,7 @@ func test_can_pause_after_yielding():
 	gut.pause_before_teardown()
 
 func test_can_call_pause_before_yielding():
+	pass_test('should  see a pause')
 	gut.pause_before_teardown()
 	gut.p('yielding for 1 second')
 	timer.start()
@@ -80,83 +88,75 @@ func test_can_call_pause_before_yielding():
 	gut.p('done yielding')
 
 func test_returning_int_does_not_cause_yield():
+	pass_test('this should not cause error')
 	return 9
 
 func test_returning_string_does_not_cause_yield():
+	pass_test('this should not cause error')
 	return 'nine'
 
 func test_returning_object_does_not_cause_yield():
-	var thing = Node2D.new()
+	pass_test('this should not cause error')
+	var thing = autofree(Node2D.new())
 	return thing
 
 func test_new_yield():
+	pass_test('should  see two 1 second pauses')
 	yield(yield_for(1, 'first yield'), 'timeout')
 	yield(yield_for(1, 'waiting around for stuff'), YIELD)
 
 func test_passing_assert_ends_yield():
 	yield(yield_for(0.5), YIELD)
-	assert_true(true)
+	pass_test('yield should stop.')
 
 func test_failing_assert_ends_yield():
 	yield(yield_for(0.5), YIELD)
-	assert_false(true, 'This should fail.')
+	fail_test('yield should stop for this failure')
 
 func test_pending_ends_yield():
 	yield(yield_for(0.5), YIELD)
 	pending('this is pending but should end test')
 
 func test_can_yield_to_signal():
-	var signaler = TimedSignaler.new()
-	add_child(signaler)
+	var signaler = add_child_autoqfree(TimedSignaler.new())
 	signaler.emit_after(.5)
 	yield(yield_to(signaler, 'the_signal', 10), YIELD)
-	assert_true(true, 'we got here')
+	pass_test('we got here')
 
 func test_after_yield_to_gut_disconnects_from_signal():
-	var signaler = TimedSignaler.new()
-	add_child(signaler)
+	var signaler = add_child_autoqfree(TimedSignaler.new())
 	signaler.emit_after(.5)
 	yield(yield_to(signaler, 'the_signal', 1), YIELD)
 	assert_false(signaler.is_connected('the_signal', gut, '_yielding_callback'))
-	remove_child(signaler)
 
 func test_yield_to__will_disconnect_after_yield_finishes_and_signal_wasnt_emitted():
-	var signaler = TimedSignaler.new()
-	add_child(signaler)
+	var signaler = add_child_autoqfree(TimedSignaler.new())
 	yield(yield_to(signaler, 'the_signal', 1), YIELD)
 	# Changing the yield to be deferred means that we have to wait again for
 	# the deferred to kick in before checking this.
 	yield(yield_for(.1), YIELD)
 	assert_false(signaler.is_connected('the_signal', gut, '_yielding_callback'))
-	remove_child(signaler)
 
 func test_yield_to__will_wait_max_time():
-	var signaler = TimedSignaler.new()
-	add_child(signaler)
+	var signaler = add_child_autoqfree(TimedSignaler.new())
 	yield(yield_to(signaler, 'the_signal', 2), YIELD)
-	assert_true(true, 'we got here')
-	remove_child(signaler)
+	pass_test('we got here')
 
 func test_yield_to__will_stop_timer_when_signal_emitted():
-	var signaler = TimedSignaler.new()
-	add_child(signaler)
+	var signaler = add_child_autoqfree(TimedSignaler.new())
 	signaler.emit_after(.5)
 	yield(yield_to(signaler, 'the_signal', 2), YIELD)
 	assert_eq(gut._yield_timer.time_left, 0.0)
-	remove_child(signaler)
 
 func test_yield_to__watches_signals():
-	var signaler = TimedSignaler.new()
-	add_child(signaler)
+	var signaler = add_child_autoqfree(TimedSignaler.new())
 	watch_signals(signaler)
 	signaler.emit_after(.5)
 	yield(yield_to(signaler, 'the_signal', 5), YIELD)
 	assert_signal_emitted(signaler, 'the_signal')
-	remove_child(signaler)
 
 func test_what_is_wrong():
-	var signaler = TimedSignaler.new()
-	add_child(signaler)
+	var signaler = add_child_autoqfree(TimedSignaler.new())
 	watch_signals(signaler)
 	signaler.emit_after(0.5)
 	yield(yield_for(1), YIELD)